change all floats to doubles

+++ src/camera.h
@@ -7,10 +7,10 @@ class camera
 public:
 	camera()
 	{
-		auto aspect_ratio = 16.0f / 9.0f;
-		auto viewport_height = 2.0f;
+		auto aspect_ratio = 16.0 / 9.0;
+		auto viewport_height = 2.0;
 		auto viewport_width = aspect_ratio * viewport_height;
-		auto focal_length = 1.0f;
+		auto focal_length = 1.0;
 
 		origin = vec3(0, 0, 0);
 		horizontal = vec3(viewport_width, 0, 0);
@@ -18,7 +18,7 @@ public:
 		lower_left_corner = origin - horizontal/2 - vertical/2 - vec3(0, 0, focal_length);
 	}
 
-	ray get_ray(const float u, const float v) const
+	ray get_ray(const double u, const double v) const
 	{
 		return ray(origin, lower_left_corner + u*horizontal + v*vertical - origin);
 	}
